+++ CPAC/resources/templates/__init__.py
@@ -1,0 +1,4 @@
+'''Template resources for C-PAC'''
+from .lookup_table import lookup_identifier
+
+__all__ = ['lookup_identifier']